Move CryptoParams to shared crypto project, make BouncyCastleCryptoParams for bouncy castle specific things
